Add resetGame function and event handler
This commit adds a new function `resetGame` to the `repository.ts` file, which resets the game by setting the card value to -1 for all players in a match. It also adds an event handler `onResetGame` in the `socket.ts` file, which calls the `resetGame` function if the user is the admin of the match.